requirements.txt
# Core ML + Data
pandas==2.2.2
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.27.5
rdkit-pypi==2024.9.3  # if installation problems, install RDKit via conda
tqdm==4.66.1

# Backend
fastapi==0.95.2
uvicorn[standard]==0.22.0

# Utils
python-multipart
